GUACAMOLE-5: Move common identifier equals() / hashCode() of abstract Identifiable implementations to common base class.
